refactor to process module
